% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_selection_utilities.R
\name{assess_subtype_with_anova}
\alias{assess_subtype_with_anova}
\title{Use ANOVA for assessing relevance of a variable to subtype}
\usage{
assess_subtype_with_anova(X)
}
\arguments{
\item{X}{Data frame with numerical data in first column, and CANCER_TYPE in another column.}
}
\value{
p-value of ANOVA test.
}
\description{
Use ANOVA for assessing relevance of a variable to subtype
}
