% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_selection_utilities.R
\name{assess_subtype_with_mglm}
\alias{assess_subtype_with_mglm}
\title{Use MGLM for assessing relevance of a variable to subtype}
\usage{
assess_subtype_with_mglm(X)
}
\arguments{
\item{X}{Data frame with numerical data in first column, and CANCER_TYPE in another column.}
}
\value{
accuracy value of fitted model.
}
\description{
Use MGLM for assessing relevance of a variable to subtype
}
