% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_selection_utilities.R
\name{assess_with_glm}
\alias{assess_with_glm}
\title{Make univariate GLM for assessing relevance of a variable}
\usage{
assess_with_glm(X, plotting = F)
}
\arguments{
\item{X}{Data frame with numerical data in first column, and SURVIVAL_STATUS in another column (filled with 0/1).}

\item{plotting}{(optional). Whether to make a plot of the ROC curve. Default FALSE.}
}
\value{
AUC of GLM for predicting the SURVIVAL_STATUS.
}
\description{
Make univariate GLM for assessing relevance of a variable
}
