% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{assess_with_ttest}
\alias{assess_with_ttest}
\title{Run t-test to assess relevance of a variable}
\usage{
assess_with_ttest(X)
}
\arguments{
\item{X}{Data frame with numerical data in first column, and SURVIVAL_STATUS in another column (filled with 0/1).}
}
\value{
p-value of t-test for the two SURVIVAL_STATUS groups.
}
\description{
Run t-test to assess relevance of a variable
}
