% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{build_and_report_final_model}
\alias{build_and_report_final_model}
\title{Build final model and report statistics}
\usage{
build_and_report_final_model(df, final_features, write, filenames)
}
\arguments{
\item{df}{The data frame.}

\item{final_features}{Features to use for final model.}

\item{write}{An identifying string for output file names.}

\item{filenames}{For reporting.}
}
\value{
cv_mean_auc Mean AUC among cross-validations in each submodel.
}
\description{
Build final model and report statistics
}
