% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{calculate_stats}
\alias{calculate_stats}
\title{Assess relevance of variables to outcome}
\usage{
calculate_stats(df, feature_names, cores = 1)
}
\arguments{
\item{df}{Data frame as returned by load_data. Rows are samples. Columns are features, including SURVIVAL_STATUS and CANCER_TYPE, as well as the input feature names.}

\item{feature_names}{Feature names (e.g. genes), as returned by load_data.}

\item{cores}{(optional) Number of cores to use for parallel computation. Default 1.}
}
\value{
stats=list(pvalue_ttest, auc_glm, pvalue_anova, acc_mglm). These four elements are vectors with one entry for each feature (with names given by feature_names).
}
\description{
Assess relevance of variables to outcome
}
