% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_selection_utilities.R
\name{create_folds_subtype}
\alias{create_folds_subtype}
\title{Cross validation fold creation respecting subtype and survival status}
\usage{
create_folds_subtype(df, k = 3)
}
\arguments{
\item{df}{Data frame as returned by load_data or pipeline. Only needs columns CANCER_TYPE and SURVIVAL_STATUS to be present.}

\item{k}{(optional) Number of parts in the partition. Default 3.}
}
\value{
A partition balanced with respect to CANCER_TYPE and SURVIVAL_STATUS
}
\description{
Cross validation fold creation respecting subtype and survival status
}
