% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{merge_rankings}
\alias{merge_rankings}
\title{Combine rankings into one}
\usage{
merge_rankings(rankings, modality = c("intersection", "union"))
}
\arguments{
\item{rankings}{Data frame or matrix. Each column is a list of elements representing a ranking. The columns should have the same elements.}

\item{modality}{Either "intersection" or "union". Corresponds to the use of the max or min function on rankings for a given named element. The effect is the same as to form the intersection (respectively union) of the first K elements from each ranking, recording the order in which new elements are added.}
}
\value{
A vector, the same length as the columns of ranking, representing the merged ranking.
}
\description{
More flexible than combining a fixed number from each.
}
