% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_selection_utilities.R
\name{merge_specific_rankings}
\alias{merge_specific_rankings}
\title{Combine t-test-based and GLM-based rankings.}
\usage{
merge_specific_rankings(
  ttest_ordered,
  glm_ordered,
  anova_ordered,
  mglm_ordered,
  feature_names
)
}
\arguments{
\item{ttest_ordered}{As returned by stats.}

\item{glm_ordered}{As returned by stats.}

\item{anova_ordered}{As returned by stats.}

\item{mglm_ordered}{As returned by stats.}

\item{feature_names}{As returned by load_data.}
}
\value{
Vector, merged ranking.
}
\description{
Combine t-test-based and GLM-based rankings.
}
