% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{rank_features}
\alias{rank_features}
\title{Sort features and write to file}
\usage{
rank_features(stats, feature_names, write = NA)
}
\arguments{
\item{stats}{As returned by calculate_stats.}

\item{feature_names}{As returned by load_data.}

\item{write}{(optional) A string added to output filenames to aid in identification (e.g. write="ge", write="cn", or write="cnge"). Default NA, no output saved to file.}
}
\value{
list(ttest_ordered, glm_ordered)
}
\description{
Utility function
}
