% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_selection_utilities.R
\name{select_representative_features}
\alias{select_representative_features}
\title{Select final representative features using hierarchical clustering}
\usage{
select_representative_features(ld, ranking, N, M)
}
\arguments{
\item{ld}{Data as returned by load_data or pipeline.}

\item{ranking}{Feature set ranking.}

\item{N}{Number of features to return.}

\item{M}{Number of features to start with.}
}
\value{
Vector of N features.
}
\description{
Select final representative features using hierarchical clustering
}
