% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{write_model_parameters_to_file}
\alias{write_model_parameters_to_file}
\title{Save model parameters}
\usage{
write_model_parameters_to_file(model_wrapper, write)
}
\arguments{
\item{model_wrapper}{A GLM as returned by build_glm.}

\item{write}{An identifying string for output file name.}
}
\description{
Save model parameters
}
