% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{write_subset}
\alias{write_subset}
\title{Write results to file}
\usage{
write_subset(features, df, write = NA)
}
\arguments{
\item{features}{Features to include.}

\item{df}{As returned by load_data.}

\item{write}{(optional) A string added to output filenames to aid in identification. Default NA, no output is saved to file.}
}
\value{
The selected dataset.
}
\description{
Write subset of data involving selected features to file. Includes CANCER_TYPE and SURVIVAL_STATUS.
}
