% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_selection_utilities.R
\name{cross_validation}
\alias{cross_validation}
\title{Perform cross validation}
\usage{
cross_validation(ld, k = 3, cores = 1, threshold = 100)
}
\arguments{
\item{ld}{Data as returned by load_data or pipeline.}

\item{k}{(optional) Number of parts in the partition. Default 3.}

\item{cores}{Number of cores for parallel computation.}

\item{threshold}{Number of features to use.}
}
\value{
...
}
\description{
Perform cross validation
}
