% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_selection_utilities.R
\name{load_data}
\alias{load_data}
\title{Load data (from brain biomarker challenge files)}
\usage{
load_data(filenames, drop.zeros = FALSE)
}
\arguments{
\item{filenames}{Vector of filenames: phenotype, outcome, and feature data.}

\item{drop.zeros}{Whether to drop features that are identically zero.}
}
\value{
list(df, feature_names). df$type, df$outcome, df$gene1, df$gene2, etc.
}
\description{
Load data (from brain biomarker challenge files)
}
