% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{pipeline}
\alias{pipeline}
\title{Pipeline}
\usage{
pipeline(
  filenames,
  ld = NULL,
  merged_ranking = NULL,
  plotting = FALSE,
  drop.zeros = FALSE,
  write = NA,
  threshold = 100,
  cores = 1
)
}
\arguments{
\item{filenames}{Filenames of the phenotype, outcome, and feature matrix (for challenges 1/2/3).}

\item{ld}{(optional). Supply the loaded data (as returned by a prior pipeline(...) call), if desired.}

\item{merged_ranking}{(optional). Supply the computed merged_ranking, to skip the pipeline forward to the last steps.}

\item{plotting}{(optional). Whether to make a plot of the ROC curves. Default FALSE.}

\item{drop.zeros}{Whether to drop features that are identically zero.}

\item{write}{(optional) A string added to output filenames to aid in identification. Default NA, no addition.}

\item{threshold}{Number of features to use in ranking stage.}

\item{cores}{Number of cores for parallel computation.}
}
\value{
ld (see load_data).
}
\description{
Pipeline
}
